% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_adjusted_hal.R
\name{make_outcome_adjusted_likelihood}
\alias{make_outcome_adjusted_likelihood}
\title{Run straitified Q fit and use the non-zero basis to restrict the basis for g fit.}
\usage{
make_outcome_adjusted_likelihood(tmle_spec, tmle_task)
}
\arguments{
\item{tmle_spec}{(tmle3_Spec) specifying the spec}

\item{tmle_task}{(tmle3_Task) specifying the data}
}
\value{
(Likelihood) initial likelihood object ready to be updated
}
\description{
Run straitified Q fit and use the non-zero basis to restrict the basis for g fit.
}
\examples{
# TO_DO
}
