% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle3_Spec_oat.R
\docType{data}
\name{tmle3_Spec_oat}
\alias{tmle3_Spec_oat}
\title{Defines a OAT TML Estimator (except for the data)}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
tmle3_Spec_oat
}
\description{
will be subclassed into TSM
}
\keyword{datasets}
